+++ src/other/assetsXml.ts
@@ -8,6 +8,10 @@ export interface IAsset {
   name?: string;
   english?: string;
   modName?: string;
+  location?: {
+    filePath: string;
+    line: number;
+  }
 }
 
 export interface IPositionedElement {
@@ -22,7 +26,7 @@ export class AssetsDocument {
 
   lines: IPositionedElement[][];
 
-  constructor(content: xmldoc.XmlDocument) {
+  constructor(content: xmldoc.XmlDocument, filePath?: string) {
     const relevantNodes = new Set<string>(['ModOps', 'ModOp', 'Asset', 'Values', 'Standard', 'GUID']);
 
     this.content = content;
@@ -33,18 +37,7 @@ export class AssetsDocument {
     while (nodeStack.length > 0) {
       const top = nodeStack.pop();
       if (top?.element.type === 'element' /*&& relevantNodes.has(top.element.name)*/) {
-        if (top.element.name === 'ModOp') {
-          // const xpath = top.element.attr['Path'];
-        }
-        
         const column = top.element.column - (top.element.position - top.element.startTagPosition + 1);
-
-        // if Property
-        if (top.history.length >= 3 && top.history[top.history.length - 3].name === 'Asset') {
-          const template = top.history[top.history.length - 3].valueWithPath('Template')
-          // current node is a property
-          let a = 1;
-        }
 
         this.getLine(top.element.line).push({ 
           history: top.history.slice(), 
@@ -55,13 +48,20 @@ export class AssetsDocument {
         if (top.element.name === 'GUID') {
           const guid = top.element.val;
           const parent = top.history.length >= 2 ? top.history[top.history.length - 2] : undefined;
+          const asset = top.history.length >= 4 ? top.history[top.history.length - 4] : undefined;
           const name = parent?.valueWithPath('Name');
   
           if (parent?.name === 'Standard' && name) {
+            const location = (filePath && asset) ? {
+              filePath,
+              line: asset?.line ?? 0
+            } : undefined;
+
             this.assets[guid] = {
               guid,
               name,
-              template: top.history.length >= 4 ? top.history[top.history.length - 4].valueWithPath('Template') : undefined
+              template: asset?.valueWithPath('Template'),
+              location
             };
           }
         }